feat: support negative pattern matching
Allow the use of a `{!:}` meta-directive to invert the matching
behaviour in tabhunter. For example, a pattern of `{!:}github.com` will
match any tab that does not contain github.com in either the title or
url.

This can also be used in combination with the existing meta-directive
for limiting matches to a given window (e.g., {w:1})